requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
pyodbc==5.2.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
